Move all infinite scrolling logic into CardDisplay
