+++ src/components/display/CardDisplay.tsx
@@ -1,7 +1,8 @@
+import { useEffect, useState } from 'react';
+import InfiniteScroll from 'react-infinite-scroll-component';
 import { Card as CardType } from "@/types/interfaces";
 import { SortMode } from "@/types/types";
 import { Link } from "react-router-dom";
-import InfiniteScroll from 'react-infinite-scroll-component';
 import CardFull from "./CardFull";
 import CardImage from "./CardImage";
 import CardList from "./CardList";
@@ -10,26 +11,45 @@ import CardText from "./CardText";
 export default function CardDisplay({
   mode,
   cards,
-  fetchMoreData,
-  hasMore
+  itemsPerPage = 20
 }: {
   mode: SortMode;
   cards: CardType[];
-  fetchMoreData: () => void;
-  hasMore: boolean;
+  itemsPerPage?: number;
 }) {
+  const [displayedCards, setDisplayedCards] = useState<CardType[]>([]);
+  const [hasMore, setHasMore] = useState(true);
+
+  useEffect(() => {
+    setDisplayedCards(cards.slice(0, itemsPerPage));
+    setHasMore(cards.length > itemsPerPage);
+  }, [cards]);
+
+  const fetchMoreData = () => {
+    const currentLength = displayedCards.length;
+    const newLength = currentLength + itemsPerPage;
+    const newCards = cards.slice(0, newLength);
+    setDisplayedCards(newCards);
+    setHasMore(newLength < cards.length);
+  };
+
   switch (mode) {
     default:
     case "image":
       return (
         <InfiniteScroll
-          dataLength={cards.length}
+          dataLength={displayedCards.length}
           next={fetchMoreData}
           hasMore={hasMore}
           loader={<h4>Loading...</h4>}
+          endMessage={
+            <p style={{ textAlign: 'center' }}>
+              <b>Yay! You have seen it all</b>
+            </p>
+          }
         >
           <div className="m-4 grid grid-cols-[repeat(auto-fill,minmax(200px,1fr))] justify-center justify-items-center gap-4">
-            {cards.map((card, index) => (
+            {displayedCards.map((card, index) => (
               <Link to={`/card/${card.attributes.card_code}`} key={index}>
                 <CardImage card={card.attributes} />
               </Link>
@@ -40,13 +60,18 @@ export default function CardDisplay({
     case "text":
       return (
         <InfiniteScroll
-          dataLength={cards.length}
+          dataLength={displayedCards.length}
           next={fetchMoreData}
           hasMore={hasMore}
           loader={<h4>Loading...</h4>}
+          endMessage={
+            <p style={{ textAlign: 'center' }}>
+              <b>Yay! You have seen it all</b>
+            </p>
+          }
         >
           <div className="m-4 grid grid-cols-[repeat(auto-fill,minmax(350px,1fr))] justify-center justify-items-center gap-2 gap-x-4 gap-y-8">
-            {cards.map((card, index) => (
+            {displayedCards.map((card, index) => (
               <Link
                 to={`/card/${card.attributes.card_code}`}
                 key={index}
@@ -61,19 +86,24 @@ export default function CardDisplay({
     case "list":
       return (
         <div className="flex overflow-x-scroll lg:justify-center">
-          <CardList cards={cards} />
+          <CardList cards={displayedCards} />
         </div>
       );
     case "full":
       return (
         <InfiniteScroll
-          dataLength={cards.length}
+          dataLength={displayedCards.length}
           next={fetchMoreData}
           hasMore={hasMore}
           loader={<h4>Loading...</h4>}
+          endMessage={
+            <p style={{ textAlign: 'center' }}>
+              <b>Yay! You have seen it all</b>
+            </p>
+          }
         >
           <div>
-            {cards.map((card, index) => (
+            {displayedCards.map((card, index) => (
               <div
                 key={index}
                 className="relative flex w-full flex-col items-center border-b border-border p-12"
